Add asynchronous JDA logic to reaction listener
Remove usage of complete() blocking method for synchronous method.
Should fix Discord heartbeat blockage.
