+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>com.misterveiga</groupId>
     <artifactId>CDS</artifactId>
-    <version>1.4.2</version>
+    <version>1.4.3</version>
     <packaging>jar</packaging>
 
 	<name>Community Discord Services</name>
